수정 : 레디스 추상화
